refactor(playground): Consolidate canary views into main path
Moves canary implementation to be the primary playground experience by:
- Moving all canary components/logic to main playground directory
- Removing mock data files
- Updating import paths across codebase
- Preserving prev/ directory for legacy code

This change completes the graduation of the canary features to production,
making them the default experience while maintaining a clean directory
structure.
