+++ app/(main)/agents/layout.tsx
@@ -5,8 +5,8 @@ import { fetchCurrentTeam } from "@/services/teams";
 import { createId } from "@paralleldrive/cuid2";
 import { redirect } from "next/navigation";
 import type { ReactNode } from "react";
-import { putGraph } from "../../(playground)/p/[agentId]/canary/actions";
-import { initGraph } from "../../(playground)/p/[agentId]/canary/lib/utils";
+import { putGraph } from "../../(playground)/p/[agentId]/actions";
+import { initGraph } from "../../(playground)/p/[agentId]/lib/utils";
 import { CreateAgentButton } from "./components";
 
 export default function Layout({
