+++ app/(playground)/p/[agentId]/components/bg.png
