+++ app/(playground)/p/[agentId]/components/header.tsx
