+++ app/(playground)/p/[agentId]/components/icons/circle-x.tsx
