+++ app/(playground)/p/[agentId]/components/icons/sparkles.tsx
