+++ app/(playground)/p/[agentId]/components/keyboard-shortcut.tsx
