+++ app/(playground)/p/[agentId]/components/node-block.tsx
