+++ app/(playground)/p/[agentId]/components/slider.tsx
