+++ app/(playground)/p/[agentId]/components/toolbar.tsx
@@ -3,8 +3,8 @@
 import * as ToggleGroup from "@radix-ui/react-toggle-group";
 import { FileUpIcon, LetterTextIcon, MousePointer2Icon } from "lucide-react";
 import { type ComponentProps, forwardRef } from "react";
-import { TextGenerationIcon } from "../../prev/beta-proto/components/icons/text-generation";
 import { useToolbar } from "../contexts/toolbar";
+import { TextGenerationIcon } from "../prev/beta-proto/components/icons/text-generation";
 import type { Tool } from "../types";
 import { Tooltip } from "./tooltip";
