+++ app/(playground)/p/[agentId]/components/viewer.tsx
@@ -2,11 +2,11 @@
 
 import * as Tabs from "@radix-ui/react-tabs";
 import { type DetailedHTMLProps, useMemo } from "react";
-import { SpinnerIcon } from "../../prev/beta-proto/components/icons/spinner";
-import { WilliIcon } from "../../prev/beta-proto/components/icons/willi";
 import { useExecution } from "../contexts/execution";
 import { useGraph } from "../contexts/graph";
 import { formatTimestamp } from "../lib/utils";
+import { SpinnerIcon } from "../prev/beta-proto/components/icons/spinner";
+import { WilliIcon } from "../prev/beta-proto/components/icons/willi";
 import type { Execution, Node, StepExecution } from "../types";
 import bg from "./bg.png";
 import ClipboardButton from "./clipboard-button";
