+++ app/(playground)/p/[agentId]/contexts/properties-panel.tsx
