+++ app/dev/copy-agent/page.tsx
@@ -1,14 +1,5 @@
-import { agents, db } from "@/drizzle";
 import { developerFlag } from "@/flags";
-import { fetchCurrentUser } from "@/services/accounts";
-import { fetchCurrentTeam } from "@/services/teams";
-import { createId } from "@paralleldrive/cuid2";
 import { notFound } from "next/navigation";
-import { putGraph } from "../../(playground)/p/[agentId]/canary/actions";
-import type {
-	AgentId,
-	Graph,
-} from "../../(playground)/p/[agentId]/canary/types";
 import AgentIdForm from "./agent-id-form";
 
 export default async function CopyAgentPage() {
